pyproject.toml
[project]
name = "medgemma-mcp"
version = "0.1.0"
description = "MCP server for MedGemma medical AI inference"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
keywords = ["mcp", "medical", "ai", "medgemma", "radiology", "healthcare"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "mcp>=1.0.0",
    "torch>=2.0.0",
    "transformers>=4.50.0",
    "accelerate>=0.20.0",
    "pillow>=9.0.0",
    "pydicom>=2.3.0",
    "pydantic>=2.0.0",
    "anyio>=4.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-anyio>=0.0.0",
    "ruff>=0.6.0",
    "pyright>=1.1.0",
]

[project.scripts]
medgemma-mcp = "medgemma_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/medgemma_mcp"]

[tool.pyright]
include = ["src/medgemma_mcp"]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
